[Backend] Exchange the file names: API, Backend
This commit exchagnes the file names of
Backend/API.ts and Backend/Backend.ts

ONE-vscode-DCO-1.0-Signed-off-by: dayo09 <dayoung.lee@samsung.com>
